requirements.txt
litellm
rich
python-dotenv
datasets